added change of mask benchmark to the examiner
Added this benchmark to be able to visualise the mask change between to
kernel launches into the same stream using the python script that
displays the SM usage. Fully functional. Can be written more cleanly
because the amount of kernels is hard coded to a set value of 2. This is
fine since it is only a test to show it is possible to change the mask
of a stream in between kernel launches.