made each mask element a class instance
This way I can free the masks in an easier way. They all save their own
index and a boolean that indicated wether they are free or not. One
single deviceProps class is instantiated that hold all the masks of the
tpcs. These masks are mask elements which I created a class for.