took care of  edge cases where amount of needed TPCs exceeded the amount of TPCs available on the device.
Split this up into two different scenarios:
1)amount needed is smaller than twice the total amount of device TPCs =>
give the job half of the total TPCs
2)amount needed is equal or greater than twice the amount of TPCs =>
give the job all of the TPCs

the way I deal with this can change in the future, especially the second
case, maybe I give it less than 100% of all the TPCs in the future.
Benchmarks will have to indicate which option to choose.